# docs/api/swagger.yaml
swagger: '2.0'
info:
  description: Entity service
  version: 1.0.0
  title: Well Delivery Service APIs
host: dns-host
basePath: /api/well-delivery
...
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  '/wellbores/v1/versions/by_well/{well_id}:actual':
    get:
      tags:
        - wellbores
      summary: Get the version number list of actual wellbore
      description: >-
        The API returns version number list of actual wellbore for the given
        well id. Required roles: service.storage.viewer, service.storage.creator
        or service.storage.admin.
      operationId: Get the version number list of actual wellbore
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: well_id
          in: path
          description: Well id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
      responses:
        '200':
          description: Actual wellbore object retrieved successfully.
          schema:
            type: array
            items:
              type: integer
              format: int64
            example:
              - 1601927059473
              - 1601927355133
              - 1601929371667
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Actual wellbore object not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Could not find related actual wellbore with well id:
                bd932928-ff2c-4c8c-9f0d-b359325b77a7
      security:
        - Bearer: []
  '/wellbores/v1/by_well/{well_id}:actual':
    get:
      tags:
        - wellbores
      summary: Get the latest version of actual wellbore
      description: >-
        The API returns latest version of actual wellbore for the given well id.
        Required roles: service.storage.viewer, service.storage.creator or
        service.storage.admin.
      operationId: Get the latest version of actual wellbore
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: well_id
          in: path
          description: Well id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
      responses:
        '200':
          description: Actual wellbore object retrieved successfully.
          schema:
            $ref: '#/definitions/EntityReturn'
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Actual wellbore object not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Could not find related actual wellbore with well id:
                bd932928-ff2c-4c8c-9f0d-b359325b77a7
      security:
        - Bearer: []
  '/wellbores/v1/by_well/{well_id}/{wellbore_version}:actual':
    get:
      tags:
        - wellbores
      summary: Get a specific version of actual wellbore
      description: >-
        The API returns specific version of actual wellbore for the given well
        id and well activity program version. Required roles:
        service.storage.viewer,  service.storage.creator or
        service.storage.admin.
      operationId: Get a specific version of actual wellbore
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: well_id
          in: path
          description: Well id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
        - name: wellbore_version
          in: path
          description: Actual wellbore version
          required: true
          type: integer
          default: 1596719439820
          format: int64
      responses:
        '200':
          description: Actual wellbore object retrieved successfully.
          schema:
            $ref: '#/definitions/EntityReturn'
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Actual wellbore object not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Could not find related actual wellbore with well id
                bd932928-ff2c-4c8c-9f0d-b359325b77a7 and and specific version
                1601570839996
      security:
        - Bearer: []
  '/wellbores/v1/versions/by_well/{well_id}:planned':
    get:
      tags:
        - wellbores
      summary: Get the version number list of planned wellbore
      description: >-
        The API returns version number list of planned wellbore for the given
        well id. Required roles: service.storage.viewer, service.storage.creator
        or service.storage.admin.
      operationId: Get the version number list of planned wellbore
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: well_id
          in: path
          description: Well id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
      responses:
        '200':
          description: Planned wellbore object retrieved successfully.
          schema:
            type: array
            items:
              type: integer
              format: int64
            example:
              - 1601927059473
              - 1601927355133
              - 1601929371667
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Planned wellbore object not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Could not find related planned wellbore with well id:
                bd932928-ff2c-4c8c-9f0d-b359325b77a7
      security:
        - Bearer: []
  '/wellbores/v1/by_well/{well_id}:planned':
    get:
      tags:
        - wellbores
      summary: Get the latest version of planned wellbore
      description: >-
        The API returns latest version of planned wellbore for the given well
        id. Required roles: service.storage.viewer, service.storage.creator or
        service.storage.admin.
      operationId: Get the latest version of planned wellbore
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: well_id
          in: path
          description: Well id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
      responses:
        '200':
          description: Planned wellbore object retrieved successfully.
          schema:
            $ref: '#/definitions/EntityReturn'
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Planned wellbore object not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Could not find related planned wellbore with well id:
                bd932928-ff2c-4c8c-9f0d-b359325b77a7
      security:
        - Bearer: []
  '/wellbores/v1/by_well/{well_id}/{wellbore_version}:planned':
    get:
      tags:
        - wellbores
      summary: Get a specific version of planned wellbore
      description: >-
        The API returns specific version of planned wellbore for the given well
        id and well activity program version. Required roles:
        service.storage.viewer,  service.storage.creator or
        service.storage.admin.
      operationId: Get a specific version of planned wellbore
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data-partition-id
          in: header
          description: This value should be the desired data partition id
          required: true
          type: string
          default: slb-osdu-dev-des-prod-testing
        - name: well_id
          in: path
          description: Well id
          required: true
          type: string
          default: 0E706AE7-8645-4BF8-9309-E1AAD1B6E54D
        - name: wellbore_version
          in: path
          description: Planned wellbore version
          required: true
          type: integer
          default: 1596719439820
          format: int64
      responses:
        '200':
          description: Planned wellbore object retrieved successfully.
          schema:
            $ref: '#/definitions/EntityReturn'
        '403':
          description: User not authorized to perform the action.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: The user is not authorized to perform this action
        '404':
          description: Planned wellbore object not found.
          schema:
            $ref: '#/definitions/AppError'
          examples:
            application/json:
              message: >-
                Could not find related planned wellbore with well id
                bd932928-ff2c-4c8c-9f0d-b359325b77a7 and and specific version
                1601570839996
      security:
        - Bearer: []
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
  AppError:
    type: object
    properties:
      message:
        type: string
...
  EntityReturn:
    type: object
    description: Entity return object
    properties:
      id:
        $ref: '#/definitions/EntityID'
      kind:
        $ref: '#/definitions/EntityKind'
      version:
        type: integer
        example: 1831253916104085
        format: int64
        description: The version number of this OSDU resource; set by the framework
      acl:
        $ref: '#/definitions/ACL'
      legal:
        $ref: '#/definitions/Legal'
      valid:
        type: boolean
        example: true
        description: The entity is valid or not.
      data:
        $ref: '#/definitions/EntityData'
...
  EntityID:
    type: string
    example: 'namespace:master-data--ExampleType:47c5967c-2961-55ea-96d7-e976a7954c26'
    description: The SRN which identifies this OSDU resource object without version.
    pattern: '^[\w\-\.]+:master-data\-\-ExampleType:[\w\-\.\:\%]+$'
  EntityKind:
    type: string
    example: 'slb-osdu-dev-des-prod-testing:well-delivery:exampletype:1.0.0'
    description: >-
      Kind of data it is being ingested and following the  pattern
      <Namespace>:<Source>:<Type>:<VersionMajor>.<VersionMinor>.<VersionPatch>. 
      The versioning scheme follows the semantic versioning,
      https://semver.org/..
    pattern: '^[\w\-\.]+:[\w\-\.]+:[\w\-\.]+:[0-9]+.[0-9]+.[0-9]+$'
  ACL:
    type: object
    description: The access control tags associated with this entity.
    required:
      - owners
      - viewers
    properties:
      viewers:
        type: array
        description: >-
          List of valid groups which will have view/read privileges over the
          entity.
        items:
          type: string
          example: data.default.viewers@slb-osdu-dev-des-prod-testing.contoso.com
      owners:
        type: array
        description: List of valid groups which will have write privileges over the entity.
        items:
          type: string
          example: data.default.owners@slb-osdu-dev-des-prod-testing.contoso.com
  Legal:
    type: object
    description: >-
      Attributes which represent the legal constraints associated with the
      record.
    properties:
      legaltags:
        type: array
        description: List of legaltag names associated with the entity.
        uniqueItems: true
        items:
          type: string
          example: sample-legal-tag
      otherRelevantDataCountries:
        type: array
        description: List of other relevant data countries.
        uniqueItems: true
        items:
          type: string
          example: US
          pattern: '^[A-Z]{2,2}$'
...
  EntityData:
    type: object
    description: Record payload represented as a list of key-value pairs.
    example:
      ExistenceKind: 'namespace:reference-data--ExistenceKind:planned:'
      NameWell: Well 01
      NameWellbore: Wellbore 01
...